+++ src/main/java/insane96mcp/survivalreimagined/mixin/PlayerMixin.java
@@ -1,7 +1,7 @@
 package insane96mcp.survivalreimagined.mixin;
 
 import insane96mcp.survivalreimagined.event.SREventFactory;
-import insane96mcp.survivalreimagined.module.combat.GoldenAbsorption;
+import insane96mcp.survivalreimagined.module.combat.RegeneratingAbsorption;
 import insane96mcp.survivalreimagined.module.experience.PlayerExperience;
 import insane96mcp.survivalreimagined.module.experience.enchantments.EnchantmentsFeature;
 import insane96mcp.survivalreimagined.module.sleeprespawn.tiredness.Tiredness;
@@ -75,7 +75,7 @@ public abstract class PlayerMixin extends LivingEntity {
 
 	@ModifyVariable(method = "actuallyHurt", at = @At(value = "STORE", ordinal = 0), ordinal = 1)
 	private float onCalculateAbsorption(float f1, DamageSource damageSource, float amount) {
-		if (GoldenAbsorption.entityAbsorption() && !(damageSource.getEntity() instanceof LivingEntity)) {
+		if (RegeneratingAbsorption.entityAbsorption() && !(damageSource.getEntity() instanceof LivingEntity)) {
 			return amount;
 		}
 		return Math.max(amount - this.getAbsorptionAmount(), 0.0F);